requirements.txt
dash==2.18.2
plotly==5.24.1
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
joblib==1.4.2
python-dateutil==2.9.0.post0
